requirements-dev.txt
# For continuous integration and development

matplotlib>=3.3.4
mock>=3.0.5
nose==1.3.7
notebook
numpy>=1.15.0,<2.0.0
Pillow
pre-commit>=20.10.0
pymongo==3.12.3
pyspark
pytest==7.2.1
scikit-learn>=0.23.2
setuptools>=47.3.1
twine>=3.2.0,<4.0.0
vulture==2.7
